2017-08-10, H, 로딩 화면 추가
