+++ image_searcher/templates/layout.html
@@ -15,8 +15,9 @@
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootswatch/3.3.7/spacelab/bootstrap.min.css"/>
     <!-- bootstrap: cosmo theme -->
     <!-- <link rel="stylesheet" href="//maxcdn.bootstrapcdn.com/bootswatch/3.3.7/cosmo/bootstrap.min.css" /> -->
-
-    <link rel=stylesheet" href="{% static "style.css" %}"/>
+    <link rel="stylesheet" href="{% static "style.css" %}"/>
+    {% block head %}
+    {% endblock %}
 </head>
 
 <body>
@@ -46,12 +47,12 @@
 
 {%  if messages %}
     <div class="container">
-    {% for message in messages %}
-        <div class="alert alert-{{ message.tags }}">
-        [{{ message.tags }}]
-        {{ message.message }}
-        </div>
-    {% endfor %}
+        {% for message in messages %}
+            <div class="alert alert-{{ message.tags }}">
+                [{{ message.tags }}]
+                {{ message.message }}
+            </div>
+        {% endfor %}
     </div>
 {% endif %}
