+++ project/static/project/style.css
@@ -1,4 +1,10 @@
 /* project/static/projcet/style.css */
-body {
-
+.preload {
+    z-index: 100;
+    display:none;
+    width:100px;
+    height: 100px;
+    position: fixed;
+    top: 50%;
+    left: 50%;
 }
