+++ project/templates/project/detail_label.html
@@ -36,7 +36,7 @@
                 <div class="modal-body">
                     <div class="alert alert-dismissible alert-success">
                         <ul>
-                            <h1><string>UPLOAD IMAGE</string></h1>
+                            <h1><strong>UPLOAD IMAGE</strong></h1>
                             <li>현재 지원하는 형식은 <strong>tar, zip, jpg, jpeg </strong>입니다.</li>
                         </ul>
                     </div>
