+++ project/templates/project/display_prediction.html
@@ -1,6 +1,58 @@
 {% extends "project/layout.html" %}
-
+{% load static %}
+{% block head %}
+    <style type="text/css">
+        .preload {
+            z-index: 100;
+            display:none;
+            width:500px;
+            height: 500px;
+            position: fixed;
+            top: 50%;
+            left: 50%;}
+        div.backLayer {
+            display:none;
+            background-color:black;
+            position:absolute;
+            opacity:0.3;
+            left:0;
+            top:0;
+            z-index: 99;
+        }
+    </style>
+    <script>
+        $(document).ready(function(){
+            $( "#submit-btn" ).click(function() {
+                var width = $(document).width();
+                var height = $(document).height();
+                $(".backLayer").width(width);
+                $(".backLayer").height(height);
+                $(".backLayer").show();
+                $(".preload").show();
+                var form = $('form')[0];
+                var formData = new FormData(form);
+                console.log("click");
+                $.ajax("{% url "project:predict" project.id %}", {
+                    method: 'POST',
+                    processData: false,
+                    contentType: false,
+                    data: formData,
+                    complete:function(xhr, settings) {
+                        $('#loading').removeAttr("disabled");
+                    },
+                    success: function (res) {
+                        $( ".preload" ).hide();
+                        console.log(res);
+                    }, error: function (xhr, status, error) {
+                        console.log(xhr);
+                    }
+                });
+            });
+        });
+    </script>
+{% endblock %}
 {% block content %}
+    <div class="backLayer" style=""></div>
     <div class="container">
         <div class="jumbotron">
             <h1><strong>IMAGE SEARCH ENGINE</strong> <sup>by KTh</sup></h1>
@@ -8,7 +60,7 @@
             <div class="col-xs-3"></div>
             <form action="{% url "project:predict" p_id=project.id %}" method=post enctype=multipart/form-data>{% csrf_token %}
                 <input type=file name=image>
-                <input type=submit value=Upload>
+                <button class="btn btn-primary" id="submit-btn">Upload</button>
             </form>
         </div>
         <div class="container-fluid">
@@ -24,4 +76,7 @@
         </div><!--/container-fluid-->
         <hr/>
         <div class="col-xs-3"><a href="{% url "project:list_label" id=project.id%}" class="btn btn-primary">돌아가기</a></div>
+    </div>
+
+    <div class="preload"><img src="{% static "project/bluecircle.gif" %}"></div>
 {% endblock %}
